the major coding and some testing has been done for the shift_predict option.
Recall: this option shifts all the-related classifiers across all bins,
hoping to align the pfd's phase with that of the trained classifiers.

Now, the predictions matrix uses the predictions from the max phase-aligned
set. We are shifting the different phases coherently...

Some thought is needed for which grid we phase-shift on (the largest or smallest),
and if taking the phase of max(Prob's) is good.

-Aaron