Aaron was generating the histograms wrong (there was an extra 'weight' factor of 100 to the prior distributions)
This commit fixes all that, making things consistent with the paper.
Note, we've coded up the weight correction in classifier.adjustscore, though it
is turned off for now. Optimal value is w=100.

As well, classifier.adjustscore accepts a 'spk' parameter, which enhances the spikes
in the prior-ratio. This defaults to 1., but is optimal at spk=1.75